+++ addons/external.go
@@ -87,6 +87,14 @@ func init() {
 		Module: "github.com/invopop/gobl.mx.cfdi",
 	})
 	tax.RegisterApprovedAddon(&tax.ExternalAddon{
+		Key: "dk-oioubl-v2",
+		Name: i18n.String{
+			i18n.EN: "Danish OIOUBL 2.1",
+			i18n.DA: "Dansk OIOUBL 2.1",
+		},
+		Module: "github.com/invopop/gobl.dk.oioubl",
+	})
+	tax.RegisterApprovedAddon(&tax.ExternalAddon{
 		Key: "fi-finvoice-v3",
 		Name: i18n.String{
 			i18n.EN: "Finland Finvoice 3.0",